Add docs input `ButtonGroup::class`.
